experiment/hungvo_experiment/requirements.txt
absl-py==2.1.0
accelerate==0.29.3
addict==2.4.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
basicsr==1.4.2
bitsandbytes==0.43.1
certifi==2024.2.2
charset-normalizer==3.3.2
cmake==3.31.0.1
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
datasets==2.18.0
deepspeed==0.14.1
diffusers==0.25.0
dill==0.3.8
einops==0.7.0
filelock==3.9.0
flatbuffers==24.3.25
fonttools==4.55.0
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
grpcio==1.68.0
hjson==3.1.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.7
imageio==2.36.0
importlib_metadata==7.1.0
Jinja2==3.1.2
kiwisolver==1.4.7
lazy_loader==0.4
lit==18.1.8
lmdb==1.5.1
Markdown==3.7
MarkupSafe==2.1.3
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu11==2.19.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
onnxruntime==1.20.1
opencv-python==4.10.0.84
packaging==24.0
pandas==2.2.2
pillow==10.2.0
platformdirs==4.3.6
protobuf==5.28.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==15.0.2
pyarrow-hotfix==0.6
pydantic==2.7.0
pydantic_core==2.18.1
pynvml==11.5.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
safetensors==0.4.3
scikit-image==0.24.0
scipy==1.14.1
six==1.16.0
sympy==1.13.1
tb-nightly==2.19.0a20241121
tensorboard-data-server==0.7.2
tifffile==2024.9.20
tokenizers==0.19.1
tomli==2.1.0
torch==2.0.1+cu118
torchvision==0.15.2+cu118
tqdm==4.66.2
transformers==4.40.0
triton==2.0.0
typing_extensions==4.8.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.1.3
xxhash==3.4.1
yapf==0.43.0
yarl==1.9.4
zipp==3.18.1
